Refactor ScoreWindow, ScoreItem and Instruments to own header
Move the classes ScoreWindow, ScoreItem and Instruments to own header.
