+++ game/instruments.hh
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "instrumentgraph.hh"
+
+#include <memory>
+#include <vector>
+
+using Instruments = std::vector<std::unique_ptr<InstrumentGraph>>;
