+++ game/scoreitem.hh
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "color.hh"
+#include "controllers.hh"
+
+struct ScoreItem {
+	int score;
+	input::DevType type;
+	std::string track;  ///< includes difficulty
+	std::string track_simple; ///< no difficulty
+	Color color;
+
+	ScoreItem() = default;
+	ScoreItem(int score, input::DevType type, std::string const& track, std::string const& track_simple, Color const& color)
+	: score(score), type(type), track(track), track_simple(track_simple), color(color) {}
+    
+	bool operator<(ScoreItem const& other) const { 
+		return score < other.score; 
+	}
+};
