+++ game/screen_sing.cc
@@ -668,101 +668,3 @@ void ScreenSing::drawMenu() {
 	m_menu.dimensions.stretch(w, h);
 }
 
-
-ScoreWindow::ScoreWindow(Instruments& instruments, Database& database):
-  m_database(database),
-  m_pos(0.8, 2.0),
-  m_bg(findFile("score_window.svg")),
-  m_scoreBar(findFile("score_bar_bg.svg"), findFile("score_bar_fg.svg"), ProgressBar::Mode::VERTICAL, 0.0, 0.0, false),
-  m_score_text(findFile("score_txt.svg")),
-  m_score_rank(findFile("score_rank.svg"))
-{
-	Game::getSingletonPtr()->showLogo();
-	m_pos.setTarget(0.0);
-	m_database.scores.clear();
-	// Singers
-	for (auto p = m_database.cur.begin(); p != m_database.cur.end();) {
-		ScoreItem item; item.type = input::DevType::VOCALS;
-		item.score = p->getScore();
-		if (item.score < 500) { p = m_database.cur.erase(p); continue; } // Dead
-		item.track = "Vocals"; // For database
-		item.track_simple = "vocals"; // For ScoreWindow
-		item.color = Color(p->m_color.r, p->m_color.g, p->m_color.b);
-
-		m_database.scores.push_back(item);
-		++p;
-	}
-	// Instruments
-	for (Instruments::iterator it = instruments.begin(); it != instruments.end();) {
-		ScoreItem item;
-		item.type = (*it)->getGraphType();
-		item.score = (*it)->getScore();
-		if (item.score < 100) { it = instruments.erase(it); continue; } // Dead
-		item.track_simple = (*it)->getTrack();
-		item.track = (*it)->getModeId();
-		item.track = UnicodeUtil::toUpper(item.track, 1); // Capitalize
-		if (item.track_simple == TrackName::DRUMS) item.color = Color(0.1, 0.1, 0.1);
-		else if (item.track_simple == TrackName::BASS) item.color = Color(0.5, 0.3, 0.1);
-		else item.color = Color(1.0, 0.0, 0.0);
-
-		m_database.scores.push_back(item);
-		++it;
-	}
-
-	if (m_database.scores.empty())
-		m_rank = _("No player!");
-	else {
-		// Determine winner
-		m_database.scores.sort([](ScoreItem i, ScoreItem j) -> bool { return (i.score>j.score); });
-		ScoreItem winner = *std::max_element(m_database.scores.begin(), m_database.scores.end());
-		int topScore = winner.score;
-		// Determine rank
-		if (winner.type == input::DevType::VOCALS) {
-			if (topScore > 8000) m_rank = _("Hit singer");
-			else if (topScore > 6000) m_rank = _("Lead singer");
-			else if (topScore > 4000) m_rank = _("Rising star");
-			else if (topScore > 2000) m_rank = _("Amateur");
-			else m_rank = _("Tone deaf");
-		} else if (winner.type == input::DevType::DANCEPAD) {
-			if (topScore > 8000) m_rank = _("Maniac");
-			else if (topScore > 6000) m_rank = _("Hoofer");
-			else if (topScore > 4000) m_rank = _("Rising star");
-			else if (topScore > 2000) m_rank = _("Amateur");
-			else m_rank = _("Loser");
-		} else {
-			if (topScore > 8000) m_rank = _("Virtuoso");
-			else if (topScore > 6000) m_rank = _("Rocker");
-			else if (topScore > 4000) m_rank = _("Rising star");
-			else if (topScore > 2000) m_rank = _("Amateur");
-			else m_rank = _("Tone deaf");
-		}
-	}
-	m_bg.dimensions.middle().center();
-}
-
-void ScoreWindow::draw() {
-	using namespace glmath;
-	Transform trans(translate(vec3(0.0, m_pos.get(), 0.0)));
-	m_bg.draw();
-	const double spacing = 0.1 + 0.1 / m_database.scores.size();
-	unsigned i = 0;
-
-	for (Database::cur_scores_t::const_iterator p = m_database.scores.begin(); p != m_database.scores.end(); ++p, ++i) {
-		int score = p->score;
-		ColorTrans c(p->color);
-		double x = spacing * (0.5 + i - 0.5 * m_database.scores.size());
-		m_scoreBar.dimensions.fixedWidth(0.09).middle(x).bottom(0.20);
-		m_scoreBar.draw(score / 10000.0);
-		m_score_text.render(std::to_string(score));
-		m_score_text.dimensions().middle(x).top(0.24).fixedHeight(0.042);
-		m_score_text.draw();
-		m_score_text.render(p->track_simple);
-		m_score_text.dimensions().middle(x).top(0.20).fixedHeight(0.042);
-		m_score_text.draw();
-	}
-	m_score_rank.draw(m_rank);
-}
-
-bool ScoreWindow::empty() {
-	return m_database.scores.empty();
-}
