+++ game/screen_sing.hh
@@ -6,10 +6,12 @@
 #include "menu.hh"
 #include "opengl_text.hh"
 #include "progressbar.hh"
+#include "scorewindow.hh"
 #include "screen.hh"
 #include "texture.hh"
 #include "theme.hh"
 #include "instrumentgraph.hh"
+#include "instruments.hh"
 
 #include <deque>
 
@@ -25,26 +27,6 @@ class ThemeInstrumentMenu;
 class ThemeSing;
 class Video;
 class Webcam;
-
-typedef std::vector<std::unique_ptr<InstrumentGraph>> Instruments;
-
-/// shows score at end of song
-class ScoreWindow {
-  public:
-	/// constructor
-	ScoreWindow(Instruments& instruments, Database& database);
-	/// draws ScoreWindow
-	void draw();
-	bool empty();
-  private:
-	Database& m_database;
-	AnimValue m_pos;
-	Texture m_bg;
-	ProgressBar m_scoreBar;
-	SvgTxtThemeSimple m_score_text;
-	SvgTxtTheme m_score_rank;
-	std::string m_rank;
-};
 
 /// class for actual singing screen
 class ScreenSing: public Screen {
